codewide: move unimplemented function stubs to IT
There exist several function implementations from `cassandra.h` in
the wrapper that just non-working stubs, not correct implementations.
This is contrary to the politics of the unimplemented functions, which
says that unimplemented functions should not be present in the
wrapper at all, in order to trigger linker errors if the user tries
to use them in their applications.

This commit removes those stubs, so that the wrapper does not pretend
to implement those functions. Users will correctly start getting
linker errors if they try to use those functions.

However, as those functions are needed for the integration tests
to compile, the stubs are not fully deleted. Instead, they are moved
to the `integration_testing` module, where they are available for
the integration tests to use, but not for the users of the wrapper.
This is because the module is only compiled if the `integration_testing`
compile flag is set.
